requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
python-dotenv==1.0.0
pyodbc==4.0.35
